requirements.txt
pyhive
thrift
thrift-sasl
python-dotenv
streamlit
pandas